+++ src/Textarea/Textarea.tsx
@@ -33,6 +33,8 @@ type BaseProps = {
   onBlur?: (e: FormEvent<HTMLTextAreaElement>) => void
   /** number of rows of input */
   rows?: number
+  /** Required flag */
+  required?: boolean
 }
 
 /** on change or on input required */
@@ -67,6 +69,7 @@ export const Textarea = forwardRef(function Textarea(
     maxLength,
     onBlur,
     rows = 4,
+    required = false,
   }: TextareaProps,
   ref: ForwardedRef<HTMLTextAreaElement>,
 ) {
@@ -78,6 +81,11 @@ export const Textarea = forwardRef(function Textarea(
           <Text tag="label" color="subtext" typo="label" htmlFor={id}>
             {label}
           </Text>
+          {required && (
+            <Text tag="label" color="error" typo="label">
+              *
+            </Text>
+          )}
         </Box>
       )}